Refactor: Replace sendGetPersonaMessage in root tsx file
